Add a working example of "asyncrifying" callbacks
That is, taking an event-based callback interface and turning it into an
asynchronous generator function. This is an attempt at distancing myself
from the Node.js way of doing anything, and a vastly-superior attempt at
achieving what Alhadis/EAL@feab5751 failed miserably at implementing.

It might wind up in the EAL library, or be powered by a function that'll
intelligently convert similar interfaces to an async generator function,
similar in spirit to Node.js's `util.promisify()`.