Implement an ugly stop-gap for transferring files
This gross, unintuitive hack is for transferring singular files across a
local hotspot, awkwardly killing the server once a POST body is received
(which is expected to be a file uploaded from an HTML form).

The messier that jg-serve(1) gets, the more I'm incentivised to complete
the rewrite started in d7fa2ba. At this point, I'm far beyond caring how
badly-written the command's implementation is, as long as its rewrite is
written correctly and cleanly.